backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
chromadb==0.5.15
sentence-transformers==3.3.1
openai==1.54.3
aiofiles==24.1.0
python-dotenv==1.0.1
numpy==1.26.4
pypdf==5.1.0
docx2python==2.0.0

